consolidate sql logic in endpoint handler
